fix: move compression.go into confighttp.go
